requirements.txt
asgiref==3.5.2
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.6
django-import-export==2.8.0
et-xmlfile==1.1.0
MarkupPy==1.14
numpy==1.23.1
odfpy==1.4.1
opencv-python==4.6.0.66
openpyxl==3.0.10
PyYAML==6.0
qrcode==7.3.1
sqlparse==0.4.2
tablib==3.2.1
xlrd==2.0.1
xlwt==1.3.0
